+++ examples/fabric/reinforcement_learning/src/agent.py
@@ -1,0 +1,255 @@
+from typing import Dict, Tuple
+
+import gymnasium as gym
+import numpy as np
+import torch
+import torch.nn.functional as F
+from src.loss import entropy_loss, policy_loss, value_loss
+from src.utils import layer_init
+from torch.distributions import Categorical
+from torchmetrics import MeanMetric
+
+from lightning.pytorch import LightningModule
+
+
+class PPOAgent(torch.nn.Module):
+    def __init__(self, envs: gym.vector.SyncVectorEnv, act_fun: str = "relu", ortho_init: bool = False) -> None:
+        super().__init__()
+        if act_fun.lower() == "relu":
+            act_fun = torch.nn.ReLU()
+        elif act_fun.lower() == "tanh":
+            act_fun = torch.nn.Tanh()
+        else:
+            raise ValueError("Unrecognized activation function: `act_fun` must be either `relu` or `tanh`")
+        self.critic = torch.nn.Sequential(
+            layer_init(
+                torch.nn.Linear(np.array(envs.single_observation_space.shape).prod(), 64), ortho_init=ortho_init
+            ),
+            act_fun,
+            layer_init(torch.nn.Linear(64, 64), ortho_init=ortho_init),
+            act_fun,
+            layer_init(torch.nn.Linear(64, 1), std=1.0, ortho_init=ortho_init),
+        )
+        self.actor = torch.nn.Sequential(
+            layer_init(
+                torch.nn.Linear(np.array(envs.single_observation_space.shape).prod(), 64), ortho_init=ortho_init
+            ),
+            act_fun,
+            layer_init(torch.nn.Linear(64, 64), ortho_init=ortho_init),
+            act_fun,
+            layer_init(torch.nn.Linear(64, envs.single_action_space.n), std=0.01, ortho_init=ortho_init),
+        )
+
+    def get_action(
+        self, x: torch.Tensor, action: torch.Tensor = None
+    ) -> Tuple[torch.Tensor, torch.Tensor, torch.Tensor]:
+        logits = self.actor(x)
+        distribution = Categorical(logits=logits)
+        if action is None:
+            action = distribution.sample()
+        return action, distribution.log_prob(action), distribution.entropy()
+
+    def get_greedy_action(self, x: torch.Tensor) -> torch.Tensor:
+        logits = self.actor(x)
+        probs = F.softmax(logits, dim=-1)
+        return torch.argmax(probs, dim=-1)
+
+    def get_value(self, x: torch.Tensor) -> torch.Tensor:
+        return self.critic(x)
+
+    def get_action_and_value(
+        self, x: torch.Tensor, action: torch.Tensor = None
+    ) -> Tuple[torch.Tensor, torch.Tensor, torch.Tensor, torch.Tensor]:
+        action, log_prob, entropy = self.get_action(x, action)
+        value = self.get_value(x)
+        return action, log_prob, entropy, value
+
+    def forward(
+        self, x: torch.Tensor, action: torch.Tensor = None
+    ) -> Tuple[torch.Tensor, torch.Tensor, torch.Tensor, torch.Tensor]:
+        return self.get_action_and_value(x, action)
+
+    @torch.no_grad()
+    def estimate_returns_and_advantages(
+        self,
+        rewards: torch.Tensor,
+        values: torch.Tensor,
+        dones: torch.Tensor,
+        next_obs: torch.Tensor,
+        next_done: torch.Tensor,
+        num_steps: int,
+        gamma: float,
+        gae_lambda: float,
+    ) -> Tuple[torch.Tensor, torch.Tensor]:
+        next_value = self.get_value(next_obs).reshape(1, -1)
+        advantages = torch.zeros_like(rewards)
+        lastgaelam = 0
+        for t in reversed(range(num_steps)):
+            if t == num_steps - 1:
+                nextnonterminal = 1.0 - next_done
+                nextvalues = next_value
+            else:
+                nextnonterminal = 1.0 - dones[t + 1]
+                nextvalues = values[t + 1]
+            delta = rewards[t] + gamma * nextvalues * nextnonterminal - values[t]
+            advantages[t] = lastgaelam = delta + gamma * gae_lambda * nextnonterminal * lastgaelam
+        returns = advantages + values
+        return returns, advantages
+
+
+class PPOLightningAgent(LightningModule):
+    def __init__(
+        self,
+        envs: gym.vector.SyncVectorEnv,
+        act_fun: str = "relu",
+        ortho_init: bool = False,
+        vf_coef: float = 0.5,
+        ent_coef: float = 0.01,
+        clip_coef: float = 0.2,
+        clip_vloss: bool = False,
+        normalize_advantages: bool = False,
+        **torchmetrics_kwargs,
+    ):
+        super().__init__()
+        if act_fun.lower() == "relu":
+            act_fun = torch.nn.ReLU()
+        elif act_fun.lower() == "tanh":
+            act_fun = torch.nn.Tanh()
+        else:
+            raise ValueError("Unrecognized activation function: `act_fun` must be either `relu` or `tanh`")
+        self.vf_coef = vf_coef
+        self.ent_coef = ent_coef
+        self.clip_coef = clip_coef
+        self.clip_vloss = clip_vloss
+        self.normalize_advantages = normalize_advantages
+        self.critic = torch.nn.Sequential(
+            layer_init(
+                torch.nn.Linear(np.array(envs.single_observation_space.shape).prod(), 64), ortho_init=ortho_init
+            ),
+            act_fun,
+            layer_init(torch.nn.Linear(64, 64), ortho_init=ortho_init),
+            act_fun,
+            layer_init(torch.nn.Linear(64, 1), std=1.0, ortho_init=ortho_init),
+        )
+        self.actor = torch.nn.Sequential(
+            layer_init(
+                torch.nn.Linear(np.array(envs.single_observation_space.shape).prod(), 64), ortho_init=ortho_init
+            ),
+            act_fun,
+            layer_init(torch.nn.Linear(64, 64), ortho_init=ortho_init),
+            act_fun,
+            layer_init(torch.nn.Linear(64, envs.single_action_space.n), std=0.01, ortho_init=ortho_init),
+        )
+        self.avg_pg_loss = MeanMetric(**torchmetrics_kwargs)
+        self.avg_value_loss = MeanMetric(**torchmetrics_kwargs)
+        self.avg_ent_loss = MeanMetric(**torchmetrics_kwargs)
+
+    def get_action(
+        self, x: torch.Tensor, action: torch.Tensor = None
+    ) -> Tuple[torch.Tensor, torch.Tensor, torch.Tensor]:
+        logits = self.actor(x)
+        distribution = Categorical(logits=logits)
+        if action is None:
+            action = distribution.sample()
+        return action, distribution.log_prob(action), distribution.entropy()
+
+    def get_greedy_action(self, x: torch.Tensor) -> torch.Tensor:
+        logits = self.actor(x)
+        probs = F.softmax(logits, dim=-1)
+        return torch.argmax(probs, dim=-1)
+
+    def get_value(self, x: torch.Tensor) -> torch.Tensor:
+        return self.critic(x)
+
+    def get_action_and_value(
+        self, x: torch.Tensor, action: torch.Tensor = None
+    ) -> Tuple[torch.Tensor, torch.Tensor, torch.Tensor, torch.Tensor]:
+        action, log_prob, entropy = self.get_action(x, action)
+        value = self.get_value(x)
+        return action, log_prob, entropy, value
+
+    def forward(
+        self, x: torch.Tensor, action: torch.Tensor = None
+    ) -> Tuple[torch.Tensor, torch.Tensor, torch.Tensor, torch.Tensor]:
+        return self.get_action_and_value(x, action)
+
+    @torch.no_grad()
+    def estimate_returns_and_advantages(
+        self,
+        rewards: torch.Tensor,
+        values: torch.Tensor,
+        dones: torch.Tensor,
+        next_obs: torch.Tensor,
+        next_done: torch.Tensor,
+        num_steps: int,
+        gamma: float,
+        gae_lambda: float,
+    ) -> Tuple[torch.Tensor, torch.Tensor]:
+        next_value = self.get_value(next_obs).reshape(1, -1)
+        advantages = torch.zeros_like(rewards)
+        lastgaelam = 0
+        for t in reversed(range(num_steps)):
+            if t == num_steps - 1:
+                nextnonterminal = 1.0 - next_done
+                nextvalues = next_value
+            else:
+                nextnonterminal = 1.0 - dones[t + 1]
+                nextvalues = values[t + 1]
+            delta = rewards[t] + gamma * nextvalues * nextnonterminal - values[t]
+            advantages[t] = lastgaelam = delta + gamma * gae_lambda * nextnonterminal * lastgaelam
+        returns = advantages + values
+        return returns, advantages
+
+    def training_step(self, batch: Dict[str, torch.Tensor]):
+        # Get actions and values given the current observations
+        _, newlogprob, entropy, newvalue = self(batch["obs"], batch["actions"].long())
+        logratio = newlogprob - batch["logprobs"]
+        ratio = logratio.exp()
+
+        # Policy loss
+        advantages = batch["advantages"]
+        if self.normalize_advantages:
+            advantages = (advantages - advantages.mean()) / (advantages.std() + 1e-8)
+
+        pg_loss = policy_loss(batch["advantages"], ratio, self.clip_coef)
+
+        # Value loss
+        v_loss = value_loss(
+            newvalue,
+            batch["values"],
+            batch["returns"],
+            self.clip_coef,
+            self.clip_vloss,
+            self.vf_coef,
+        )
+
+        # Entropy loss
+        ent_loss = entropy_loss(entropy, self.ent_coef)
+
+        # Update metrics
+        self.avg_pg_loss(pg_loss)
+        self.avg_value_loss(v_loss)
+        self.avg_ent_loss(ent_loss)
+
+        # Overall loss
+        return pg_loss + ent_loss + v_loss
+
+    def on_train_epoch_end(self, global_step: int) -> None:
+        # Log metrics and reset their internal state
+        self.logger.log_metrics(
+            {
+                "Loss/policy_loss": self.avg_pg_loss.compute(),
+                "Loss/value_loss": self.avg_value_loss.compute(),
+                "Loss/entropy_loss": self.avg_ent_loss.compute(),
+            },
+            global_step,
+        )
+        self.reset_metrics()
+
+    def reset_metrics(self):
+        self.avg_pg_loss.reset()
+        self.avg_value_loss.reset()
+        self.avg_ent_loss.reset()
+
+    def configure_optimizers(self, lr: float):
+        return torch.optim.Adam(self.parameters(), lr=lr, eps=1e-4)
